scheduler_refactored/pyproject.toml
[tool.poetry]
name = "scheduler-ms"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pika="^1.3.1"
python-crontab = "^2.7.1"
python-dotenv = "^1.0.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
